+++ .github/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 <h3>New features since last release</h3>
 
-* The Hamiltonian can now store grouping information, which can be accessed by a devise to 
+* The Hamiltonian can now store grouping information, which can be accessed by a device to 
   speed up computations of the expectation of a Hamiltonian. 
   [(#1515)](https://github.com/PennyLaneAI/pennylane/pull/1515)
 
@@ -11,7 +11,7 @@
   coeffs = np.array([1., 2., 3.])
   H = qml.Hamiltonian(coeffs, obs, compute_groupings=True)
   ```
-  This initialization stores the indices needed to make groups of 
+  Initialization with ``compute_groupings=True`` stores the indices required to make groups of 
   commuting observables and their coefficients. These are used 
   (or, if not found, computed) by the ``get_groupings()`` method, which returns the 
   actual groupings.